Adding python notebooks used to produce figures and results
These are the notebook used for the majority of the work in the PINK paper

Note that for each of use, a number of symbolic links were created to easily reference appropriate directories. Be mindful of this. You will also have to ensure that you have the SOMs from PINK in a somewhat consistent manner to what I have. The use of the pink_utils should make this fairly trival though to simply change teh directoy/path names